Refactor PTSBE to avoid leaking nvqir headers in public include chain.
Split PTSBESampler.h into public header (no nvqir dependencies) and
internal PTSBESamplerImpl.h. Move template bodies to PTSBESampler.cpp
with explicit instantiations for float and double. Remove unnecessary
cudaq/simulators.h include from PTSBESampleIntegration.h.
